VACMS-15414 Move Cypress tests into megamenu and footer folders
