+++ src/platform/utilities/tests/header-footer/utilities/helpers.js
